Mingo game
There is a game they termed as Mingo. A random generator (like a
speaker standing in a group house game calls out a number) generates
out any number from 1 to 1000.There is a 10X10 matrix. A random
generator assigns values to each block of this matrix(within 1 to 1000
obviously). Whenever, a row or a column or a diagonal is fully filled
in this 10x10 from the numbers called out by the speaker, its called a
'Mingo'. Write a program that will find first Mingo, then second Mingo,
then thirds Mingo.…nd so forth.